### 2019-07-02 #1540 @whedon
[REVIEW]: Sourcepredict: Prediction of metagenomic sample sources using dimension reduction followed by machine learning classification
**Submitting author:** @maxibor (<a href="http://orcid.org/0000-0001-9140-7559">Maxime Borry</a>)
**Repository:** <a href="https://github.com/maxibor/sourcepredict" target ="_blank">https://github.com/maxibor/sourcepredict</a>
**Version:** v0.34
