### 2019-07-04 0f75504
Update install instructions
