+++ docs/install.md
@@ -2,4 +2,4 @@
 
 SourcePredict can be installed using the [conda](https://conda.io/en/latest/miniconda.html) package manager:
 
-`$ conda install -c etetoolkit -c bioconda -c maxibor sourcepredict`
+`$ conda install -c conda-forge -c etetoolkit -c bioconda -c maxibor sourcepredict`
