### 2019-07-04 maxibor/sourcepredict#2 @maxibor
Hi @gavinmdouglas,
First of all, thanks for taking the time to review SourcePredict for JOSS :)

There was indeed a conda channel missing in the installation intructions, it should be fixed by now with 0f75504.
The new install instructions are `conda install -c conda-forge -c etetoolkit -c bioconda -c maxibor sourcepredict`
