### 2019-07-04 maxibor/sourcepredict#2 @gavinmdouglas
v0.33 is now installed correctly and the example command works.
