### 2019-07-09 fb5ac07
fix lowercase in normalization variable name
