### 2019-07-09 maxibor/sourcepredict#3 @gavinmdouglas
JOSS review
Related to https://github.com/openjournals/joss-reviews/issues/1540

I think sourcepredict is really interesting and could potentially be very useful for metagenomics researchers. However, I think several areas of the documentation need to be clarified.

## Major comments:

* I think the key question users will have is which taxonomic classifier should be used. Different classifiers can result in starkly different results and so from the user's perspective it's very unclear how  the use of different classifiers will affect the result. If you have only tested Kraken1 with sourcepredict I think this should be clarified. Either way I think it should be emphasized that exactly the same taxonomic classifier should be used for the reference sources as for the sink samples (if you disagree than this should definitely be clarified).

* Would you recommend users use the default database or do you envision users plugging in custom reference files? The latter makes more sense to me since the combination of human, dog, and stool doesn't seem very generalizable. If you agree than more documentation on how to construct custom reference files is needed and adding a sentence in the manuscript related to this point would be useful.

* A more real-world usage example is needed since I think having the "ERR1915662_copy" sample is confusing. Ideally an example of how to identify an outlier contaminated sample or how to discriminate 1 dog and 1 human sample would be more informative. A couple of sentences recommending users how to interpret the output and what next steps to take would also be useful (e.g. do you have any recommendations for hard cut-offs for saying that a sample is sufficiently of a single source? If not then this would be good to clarify since I'm sure users will be trying to figure that out).

* No performance claims were made in the manuscript so I think it's beyond the scope of my review to request performance validations. However, I think some sort of validation of the accuracy of the tool or comparison with a tool like SourceTracker would be really useful and probably increase usage of the tool.

## README comments:

* Remove extra `wget` typo in README dog example

* The README example would benefit from a very brief description of what the input file is, what is being performed, and what the output file means in addition to the rough expected runtime.

* I suggest that the intermediate log output be removed from the README for ease of reading

* Outputting the tSNE embedding and showing an example plot would also be really informative for users looking at the test example.

* Possible to add brief description of how unittests can be run or are these run automatically by conda? (checkbox: `Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?`)

* Community guidelines needed (checkbox: `Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support`)

## Manuscript edits / comments

### General

- Adding a sentence that gives a better motivation for why source detection is useful is needed. Can you give an example of when the source of a sample is unknown? Also, I think many users would be interested in identifying outlier contaminated samples using this approach even if the sampling source is known. If you think the method would be appropriate for these cases I think it would help motivate the need for sourcepredict to bring up that example.

- Also, adding a sentence or two mentioning what the reference sources are and what organisms they're based on is needed. Also need to mention what the reference tree is for the UniFrac calculation.

- Spacing between paragraphs is inconsistent - I'm not sure if this is due to the PDF creation or in the original document.

- I think many of the paragraphs thoughout that are single sentences could be combined. For instance, the paragraph starting with "One aspect of metagenomics" should be combined with preceding paragraph.

- Missing DOIs for 2 citations - are these available?

### Summary section

- "a Python Conda package" sounds odd to me; is it more accurate to say "a Python package distributed through Conda"?

- "**will** compute the organism taxonomic..." --> "compute the organism taxonomic..."

- The sentence starting with "These taxonomic classifiers" is essentially repeating the preceding sentence (e.g. "community composition of organisms" vs "organism taxonomic composition") and giving Kraken as an example. I think this should be re-worded to make it less redundant.

- I suggest the last sentence of the summary be re-worded to make it more clear, something like:
"However, the Sourcepredict results are easier interpreted since the samples are embedded in a human observable low-dimensional space. This embedding is performed by a dimension reduction algorithm followed by K-Nearest-Neighbours (KNN) classification."

### Method section

- Write out what GMPR stands for on first usage

- I suggest using the term "source" throughout in place of "class" to avoid confusion. They are currently used interchangeably.

- Would it be more accurate to call the "estimated samples" "simulated samples" instead?
