### 2019-07-10 23d516a
change python conda package formulation
