### 2019-07-10 85977cf
simplify formulation
