### 2019-07-10 4c97a30
better summary closing sentence
