### 2019-07-10 d3cdf5a
update method section after review 1
