### 2019-07-10 a56f524
remove extra wget and add data origin
