### 2019-07-10 c8ecb70
update intro to match paper
