### 2019-07-10 6ff5892
update usage with sink file example
