### 2019-07-10 fb5e085
add links to doc for input files
