### 2019-07-10 e3fdac4
update example usage
