### 2019-07-10 4115f40
update readme example with output
