+++ docs/index.rst
@@ -20,6 +20,7 @@ __ homepage_
    usage
    results
    methods
+   custom_sources
 
 
 
