### 2019-07-10 25c3079
remove data from conda and bump ver to 0.34
