+++ conda/build.sh
@@ -2,7 +2,5 @@
 
 cp sourcepredict $PREFIX/bin
 cp -r sourcepredictlib $PREFIX/bin
-mkdir -p $PREFIX/bin/data
-cp data/*.csv $PREFIX/bin/data
 
 
