### 2019-07-10 8c3a9b7
add examples in doc
