+++ docs/run_example.md
@@ -1,0 +1,546 @@
+
+# Sourcepredict analysis example
+
+
+```python
+import pandas as pd
+```
+
+In this example, we'll apply Sourcepredict to the example data.   
+
+The [example source files](https://github.com/maxibor/sourcepredict/blob/master/data/modern_gut_microbiomes_sources.csv) contains the following sources:
+- *Homo sapiens* gut microbiome ([1](https://doi.org/10.1038/nature11234), [2](https://doi.org/10.1093/gigascience/giz004), [3](https://doi.org/10.1038/s41564-019-0409-6), [4](https://doi.org/10.1016/j.cell.2019.01.001), [5](https://doi.org/10.1038/ncomms7505), [6](http://doi.org/10.1016/j.cub.2015.04.055))
+- *Canis familiaris* gut microbiome ([1](https://doi.org/10.1186/s40168-018-0450-3))
+- Soil microbiome ([1](https://doi.org/10.1073/pnas.1215210110), [2](https://www.ncbi.nlm.nih.gov/bioproject/?term=322597), [3](https://dx.doi.org/10.1128%2FAEM.01646-17))
+
+
+```python
+src = pd.read_csv("../data/modern_gut_microbiomes_sources.csv", index_col=0)
+labels = pd.read_csv("../data/modern_gut_microbiomes_labels.csv", index_col=0)
+```
+
+A source file is a OTU count table containing the samples as columns headers, and the TAXID as row indices
+
+
+```python
+src.head()
+```
+
+
+
+
+<div>
+<style scoped>
+    .dataframe tbody tr th:only-of-type {
+        vertical-align: middle;
+    }
+
+    .dataframe tbody tr th {
+        vertical-align: top;
+    }
+
+    .dataframe thead th {
+        text-align: right;
+    }
+</style>
+<table border="1" class="dataframe">
+  <thead>
+    <tr style="text-align: right;">
+      <th></th>
+      <th>SRR1175007</th>
+      <th>SRR042182</th>
+      <th>SRR061154</th>
+      <th>SRR061499</th>
+      <th>SRR063469</th>
+      <th>SRR062324</th>
+      <th>SRR1179037</th>
+      <th>SRR061236</th>
+      <th>SRR061456</th>
+      <th>SRR642021</th>
+      <th>...</th>
+      <th>mgm4477903_3</th>
+      <th>mgm4477807_3</th>
+      <th>mgm4477874_3</th>
+      <th>mgm4477904_3</th>
+      <th>mgm4477804_3</th>
+      <th>mgm4477873_3</th>
+      <th>ERR1939166</th>
+      <th>SRR3578625</th>
+      <th>ERR1939165</th>
+      <th>SRR3578645</th>
+    </tr>
+    <tr>
+      <th>TAXID</th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+    </tr>
+  </thead>
+  <tbody>
+    <tr>
+      <th>0</th>
+      <td>3528337.0</td>
+      <td>11563613.0</td>
+      <td>10084261.0</td>
+      <td>20054993.0</td>
+      <td>8747525.0</td>
+      <td>12116517.0</td>
+      <td>4191329.0</td>
+      <td>13992760.0</td>
+      <td>14825759.0</td>
[…]
