### 2019-07-10 44e7f81
result interpretation
