### 2019-07-10 e6380f5
Update issue templates
