+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,23 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: "[BUG]"
+labels: ''
+assignees: ''
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**To Reproduce**
+Steps to reproduce the behavior:
+
+**Sourcepredict version**
+Version: [e.g. 0.33]
+
+**OS**
+ OS: [e.g. Ubuntu 18.04]
+
+**Additional context**
+Add any other context about the problem here.
