### 2019-07-10 108fb0b
Update issue templates
