### 2019-07-15 #1540 @gavinmdouglas
Thanks @lpantano - I made comments (see: https://github.com/maxibor/sourcepredict/issues/3), which I believe are still being addressed by the author. I'll post here when that issue is resolved.
