### 2019-07-18 97cfeea
add "no normalization" norm
