### 2019-07-18 5002619
add "no norm" and change bray-curtis to all taxid
