+++ docs/run_example.md
@@ -1,29 +1,31 @@
 
-# Sourcepredict analysis example
+# Sourcepredict example 1: Gut host species prediction
 
 
 ```python
 import pandas as pd
 ```
 
-In this example, we'll apply Sourcepredict to the example data.   
+In this example, we'll apply Sourcepredict to the example dataset provided in the Sourcepredict directory.   
 
-The [example source files](https://github.com/maxibor/sourcepredict/blob/master/data/modern_gut_microbiomes_sources.csv) contains the following sources:
+The [example datasets](https://github.com/maxibor/sourcepredict/blob/master/data/modern_gut_microbiomes_sources.csv) contains the following samples:
 - *Homo sapiens* gut microbiome ([1](https://doi.org/10.1038/nature11234), [2](https://doi.org/10.1093/gigascience/giz004), [3](https://doi.org/10.1038/s41564-019-0409-6), [4](https://doi.org/10.1016/j.cell.2019.01.001), [5](https://doi.org/10.1038/ncomms7505), [6](http://doi.org/10.1016/j.cub.2015.04.055))
 - *Canis familiaris* gut microbiome ([1](https://doi.org/10.1186/s40168-018-0450-3))
 - Soil microbiome ([1](https://doi.org/10.1073/pnas.1215210110), [2](https://www.ncbi.nlm.nih.gov/bioproject/?term=322597), [3](https://dx.doi.org/10.1128%2FAEM.01646-17))
 
+## Preparing the data
+
 
 ```python
-src = pd.read_csv("../data/modern_gut_microbiomes_sources.csv", index_col=0)
+otus = pd.read_csv("../data/modern_gut_microbiomes_sources.csv", index_col=0)
 labels = pd.read_csv("../data/modern_gut_microbiomes_labels.csv", index_col=0)
 ```
 
-A source file is a OTU count table containing the samples as columns headers, and the TAXID as row indices
+This is a OTU count table containing the samples as columns headers, and the TAXID as row indices
 
 
 ```python
-src.head()
+otus.head()
 ```
 
 
@@ -222,7 +224,7 @@ src.head()
 
 
 
-The labels file contains the mapping of samples names with their actual sources
+The labels file contains the mapping of samples names with their actual origin (sources)
 
 
 ```python
@@ -280,84 +282,59 @@ labels.head()
 
 
 
-Here, we'll use a sink file coming for a [dog gut microbiome study](https://microbiomejournal.biomedcentral.com/articles/10.1186/s40168-018-0450-3).  
-A sink file is formatted exactly like a source file, with samples as columns headers, and the TAXID as row indices.
+We will divide the source in training (95%) and testing (5%) dataset
 
 
 ```python
-sink = pd.read_csv("../data/test/dog_test_sample.csv", index_col=0)
-sink.head()
+otus_train = otus.sample(frac=0.95, axis=1)
+otus_test = otus.drop(otus_train.columns, axis=1)
 ```
 
+We also have to subset the labels file to only the training dataset 
 
 
+```python
+train_labels = labels.loc[otus_train.columns,:]
+test_labels = labels.loc[otus_test.columns,:]
+```
 
-<div>
-<style scoped>
-    .dataframe tbody tr th:only-of-type {
-        vertical-align: middle;
-    }
+## Sourcepredict
 
-    .dataframe tbody tr th {
-        vertical-align: top;
[…]
