### 2019-07-18 9e35118
[skip travis] update paper
