### 2019-07-19 maxibor/sourcepredict#3 @gavinmdouglas
My major comments have been addressed, but I do have a few remaining minor points which I think should be addressed:

* The example file is referred to as an OTU table, but this is actually the output of kraken right? OTU typically just refers to clustered amplicon reads (e.g. based on 16S) so this should be re-phrased or at the very least clarified what you mean by "OTU table" in the example (i.e. a taxonomic abundance table).

* I think it's fine to write up the paper as a series of very short paragraphs, but the paragraphs should be denoted consistently. It's currently a little hard to read since the spacing is inconsistent - I think there should either be an indentation at the start of each paragraph or a blank line separating them.

* Typo in documentation: "the taxonomic classifier used to produce the source OTU count table must **be** the same as the one used to produced the sink OTU count table."

* Manuscript typo: "Sourcepredict results are **more easily** interpreted..."
