### 2019-07-19 #1540 @gavinmdouglas
Hey @lpantano - There a small number of minor changes to the documentation that I think should be made, but all of my major comments have been addressed. Note that DOIs are not available for two of the references apparently.
