### 2019-07-19 #1540 @lpantano
Thank you, @gavinmdouglas, for all the work. @maxibor, are you willing to make the changes to the documentation as the reviewer suggest?
