### 2019-07-19 #1540 @lpantano
@whedone remove @giraola as reviewer
