### 2019-07-26 b59eb73
expose knn params to user
