+++ tests/test_utils.py
@@ -16,6 +16,7 @@ def test_checks():
     assert utils.check_norm('None') == "no normalization"
     assert utils.check_embed('tsne') == 'TSNE'
     assert utils.check_distance('Weighted_unifrac') == 'weighted_unifrac'
+    assert utils.check_weigths('Uniform') == 'uniform'
     assert utils.check_gen_seed(42) == 42
     assert type(utils.check_gen_seed(seed=None)) is int
 
