### 2019-07-26 78a5858
update doc with more examples and mixed prop
