+++ docs/requirements.txt
@@ -1,0 +1,3 @@
+ipykernel
+nbsphinx
+sphinxcontrib-napoleon
