### 2019-07-26 maxibor/sourcepredict#3 @gavinmdouglas
Great, all of my comments have been addressed!
