### 2019-07-26 #1540 @gavinmdouglas
Hey @lpantano  - quick note that all of my remaining minor comments have been addressed.
