### 2019-07-26 maxibor/sourcepredict#3 @maxibor
Thanks a lot @gavinmdouglas for all the comments and suggestions !
