### 2019-08-21 #1540 @will-rowe
Happy to review if needed.
