### 2019-08-27 #1540 @will-rowe
Hi all. I've made a few minor comments over at the [sourcepredict issue tracker](https://github.com/maxibor/sourcepredict/issues/4). Nothing major - it's a nice contribution!
