### 2019-08-27 bcaf000
rename dog test sink sample
