### 2019-08-27 185136a
fix ref formatting in paper
