### 2019-08-27 maxibor/sourcepredict#4 @maxibor
Thanks for the review @will-rowe !

Here are my answers and modifications based on your comments.

## Minor comments

> File format descriptor for the taxonomic classification tables
> You use several terms to refer to the same thing throughout the documentation, such as OTU count table, TAXID abundance count table and abundance table. Please unify these.

Fixed with 80dbff9

> Input filename in example
> Rather than calling the example input data `dog_example.csv`, could you make it a more informative filename? Even something like `dog_sink_sample.csv`, to tie it in with your source-sink narrative.

Changed with bcaf000

> Reference formatting issue in the paper
> The formatting for the Kraken reference in the second paragraph of the paper needs fixing, at the moment the authors are outside the reference parenthesise.

Nice catch, I didn't see this one. Fixed with 185136a

## Suggestions

> Continuity in syntax/documentation
> To be clear, I don't need to see this changed but I just wanted to raise it. I found the usage documentation a little confusing when you used the term `abundance_table`, for instance as the positional argument tag. Throughout the paper and documentation you have nicely set up the source-to-sink logic and I think it is a shame not have this continuity here.

Improved with 80dbff9

> Standardised format for input taxonomic classification tables
> Again, just a suggestion for a future release maybe. It may be worth considering a standardised format, e.g. biom format. I only suggest this as the parsing required to get the TAXID and abundance values from different taxonomic classifiers may make this tool less appealing or unaccessible to some users. If you could get the standard output from one or more taxonomic classifiers (kraken/kaiju/metaphlan etc.) and run them straight into sourcepredict, that would lower the barrier to entry. Your [kraken pipeline](https://github.com/maxibor/kraken-nf) is a great step in this direction. You could maybe think about adding the kraken_parse script to this repo too - particularly as people often receive kraken reports as part of sequencing results so may not want/need to re-run kraken (or use mini kraken) before using sourcepredict.

Getting rid of the NCBI TAXID is definitely something I'm thinking about. For the moment, they are needed to run the unifrac distance method, but I may work on a more general version of sourcepredict, not necessarily based on taxonomics identifiers or species, in the future.

I added a link to the `kraken_parse.py` script in the documentation 335d03f
