### 2019-08-27 maxibor/sourcepredict#4 @will-rowe
Awesome - fast work! I'm happy and will close the issue for the review
