### 2019-08-28 #1540 @maxibor
@lpantano I created the Zenodo archive and added DOI badge in the README.
The version is now 0.34 (following all the modifications after this review).
The rest should be fine 😃
