### 2019-08-28 89b5cdf
update conda recipe for v0.34
