+++ conda/build.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 
+mkdir -p $PREFIX/bin
 cp sourcepredict $PREFIX/bin
 cp -r sourcepredictlib $PREFIX/bin
 
