+++ conda/meta.yaml
@@ -3,7 +3,7 @@ package:
   version: "0.34"
 
 source:
-  git_rev: "0.34"
+  git_rev: 0.34
   git_url: https://github.com/maxibor/sourcepredict.git
 
 requirements:
