### 2019-08-29 #1540 @labarba
Some editorial fixes/requests:

- Remove the URL to the GitHub repository from the first line. The JOSS article template includes the repo as part of the margin decorators on the front page.
- Add comma after "i.e." and don't italicize
- [p.1] "Second it predicts" --> add comma
- [p.2] heading "Prediction of unknown sources proportion" --> Prediction of the proportion of unknown sources
- "Let S_i ... be a sample ... be an organism" --> pls fix verb repeat
- "of each of the ... organism" --> organisms (plural), or remove "of the"
- "is added to each U_k^S samples --> here, it's "sample" in singular: each sample
- independantly --> independently
- "Prediction of known source proportion" --> Prediction of the proportion of known sources
- [p.3] "Combining unknown and source proportion" --> proportions (plural)
- "Then for each" --> delete "Then" (you've started a new section); appears twice
- Acknowledgements --> add a \ after "Dr." to suppress the double white space (default in LaTeX for end of sentence)
- Bibliography: add case protection in titles with `{}` to keep the capitalization, where appropriate; for example, in "Southern Wisconsin"
- Use camel case for journal titles
- DOI for Maaten & Hinton 2008 (and caps protection)
- DOI for Patt et al. 1999?

You seem to have written the sections on p,2 as a sequence of orphan sentences. Can you group into paragraphs?

You mention `scikit-bio`: is there a reference you can cite for this library? What about `scikit-learn`?
