+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @article{kraken,
   title={Kraken: ultrafast metagenomic sequence classification using exact alignments},
   author={Wood, Derrick E and Salzberg, Steven L},
-  journal={Genome biology},
+  journal={Genome Biology},
   volume={15},
   number={3},
   pages={R46},
@@ -31,9 +31,9 @@
   doi={10.1038/455481a}
 }
 @article{scikit-learn,
-  title={Scikit-learn: Machine learning in Python},
+  title={{Scikit-learn: Machine learning in Python}},
   author={Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and others},
-  journal={Journal of machine learning research},
+  journal={Journal of Machine Learning Research},
   volume={12},
   number={Oct},
   pages={2825--2830},
@@ -42,7 +42,7 @@
 @article{platt,
   title={Probabilistic outputs for support vector machines and comparisons to regularized likelihood methods},
   author={Platt, John and others},
-  journal={Advances in large margin classifiers},
+  journal={Advances in Large Margin Classifiers},
   volume={10},
   number={3},
   pages={61--74},
@@ -52,7 +52,7 @@
 @article{ete3,
   title={ETE 3: reconstruction, analysis, and visualization of phylogenomic data},
   author={Huerta-Cepas, Jaime and Serra, Fran{\c{c}}ois and Bork, Peer},
-  journal={Molecular biology and evolution},
+  journal={Molecular Biology and Evolution},
   volume={33},
   number={6},
   pages={1635--1638},
@@ -72,9 +72,9 @@
   doi={10.1128/AEM.01996-06}
 }
 @article{tsne,
-  title={Visualizing data using t-SNE},
+  title={Visualizing data using {t-SNE}},
   author={Maaten, Laurens van der and Hinton, Geoffrey},
-  journal={Journal of machine learning research},
+  journal={Journal of Machine Learning Research},
   volume={9},
   number={Nov},
   pages={2579--2605},
@@ -92,7 +92,7 @@
   doi={10.1038/nmeth.1650}
 }
 @article{bray-curtis,
-  title={An ordination of the upland forest communities of southern Wisconsin},
+  title={An ordination of the upland forest communities of {southern Wisconsin}},
   author={Bray, J Roger and Curtis, John T},
   journal={Ecological monographs},
   volume={27},
@@ -101,6 +101,44 @@
   year={1957},
   publisher={Wiley Online Library},
   doi={10.2307/1942268}
+}
+@misc{scikit-bio,
+  author       = {Jai Ram Rideout and
+                  Greg Caporaso and
+                  Evan Bolyen and
+                  Daniel McDonald and
+                  Yoshiki Vázquez Baeza and
+                  Jorge Cañardo Alastuey and
+                  Anders Pitman and
+                  Jamie Morton and
+                  Jose Navas and
+                  Kestrel Gorlick and
+                  Justine Debelius and
+                  Zech Xu and
+                  llcooljohn and
+                  adamrp and
+                  Joshua Shorenstein and
+                  Laurent Luce and
+                  Will Van Treuren and
+                  John Chase and
+                  charudatta-navare and
+                  Colin Brislawn and
+                  Antonio Gonzalez and
+                  Weronika Patena and
+                  Karen Schwarzberg and
+                  teravest and
+                  Jens Reeder and
+                  shiffer1 and
+                  nbresnick and
+                  Kevin Murray and
+                  alexbrc and
[…]
