### 2019-08-30 #1540 @maxibor
Thanks for the editorial check and comments !

> Remove the URL to the GitHub repository from the first line. The JOSS article template includes the repo as part of the margin decorators on the front page.

Done

> Add comma after "i.e." and don't italicize

Done

> [p.1] "Second it predicts" --> add comma

Done

> [p.2] heading "Prediction of unknown sources proportion" --> Prediction of the proportion of unknown sources

Done

> "Let S_i ... be a sample ... be an organism" --> pls fix verb repeat

Done

> "of each of the ... organism" --> organisms (plural), or remove "of the"

Done

> "is added to each U_k^S samples --> here, it's "sample" in singular: each sample

Done

> independantly --> independently

Done

> "Prediction of known source proportion" --> Prediction of the proportion of known sources

Done

> [p.3] "Combining unknown and source proportion" --> proportions (plural)

Done

> "Then for each" --> delete "Then" (you've started a new section); appears twice

Done

> Acknowledgements --> add a \ after "Dr." to suppress the double white space (default in LaTeX for end of sentence)

Done

> Bibliography: add case protection in titles with `{}`to keep the capitalization, where appropriate; for example, in "Southern Wisconsin"

Done

> Use camel case for journal titles

Done

> DOI for Maaten & Hinton 2008 (and caps protection)

> DOI for Patt et al. 1999?

As far as I could search, there is no DOI available for these articles (I couldn't find any DOI for the Journal of Machine Learning research)

> You seem to have written the sections on p,2 as a sequence of orphan sentences. Can you group into paragraphs?

Done

> You mention `scikit-bio`: is there a reference you can cite for this library?

Added a reference

> What about `scikit-learn`?

Added a reference
