### 2019-09-04 #1540 @labarba
Big thanks to our editor: @lpantano, and reviewers: @gavinmdouglas, @will-rowe — we appreciate your contributions to JOSS! 🙏
