### 2019-09-04 #1540 @arfon
Looks like the DOI is resolving now.
