### 2019-09-04 #1540 @maxibor
Thanks @lpantano and @labarba for the editing !
Thanks a lot @gavinmdouglas and @will-rowe for the review and really helpful comments and suggestions !
First paper of my PhD, and great publishing experience with JOSS !
