### 2019-09-04 #1540 @arfon
> First paper of my PhD, and great publishing experience with JOSS !

🎉 congrats!
